requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
PyMuPDF>=1.24.0
reportlab>=4.2.0
pdf2image>=1.17.0
Pillow>=10.3.0
openai>=1.12.0
anthropic>=0.39.0
google-generativeai>=0.8.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
httpx>=0.28.0
